+++ oshi-common/src/main/java/module-info.java
@@ -5,12 +5,15 @@
 module com.github.oshi.common {
     exports oshi.annotation;
     exports oshi.annotation.concurrent;
+    exports oshi.driver.common.windows.perfmon;
     exports oshi.hardware;
     exports oshi.hardware.common;
     exports oshi.hardware.common.platform.linux;
     exports oshi.hardware.common.platform.unix;
+    exports oshi.hardware.common.platform.windows;
     exports oshi.software.common;
     exports oshi.software.common.os.linux;
+    exports oshi.software.common.os.windows;
     exports oshi.software.os;
     exports oshi.util;
     exports oshi.util.driver.linux;