+++ oshi-common/src/main/java/oshi/driver/common/windows/perfmon/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Provides common Windows performance monitor constants and definitions shared between JNA and FFM implementations.
+ */
+package oshi.driver.common.windows.perfmon;