+++ oshi-common/src/main/java/oshi/software/common/os/windows/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Provides common implementations for Windows operating system information shared between JNA and FFM implementations.
+ */
+package oshi.software.common.os.windows;